5: added prefix team in component page + added is in variable
